Chat-Bot Coursework/requirements.txt
requests~=2.28.1
pony~=0.7.16
Pillow~=9.2.0